adjust misc contract deployments
