+++ protocol/reseed/dataConverts/convertWhitelist.js
@@ -40,13 +40,13 @@ const tokenToOracleMapping = {
     "0x000000000000000000000000000000000000000000000000000000000001fa40"
   ],
   "0x5979D7b546E38E414F7E9822514be443A4800529": [
-    "0x74749a1c8b2faa03c7259F90919628b1241A2ea5",
+    "0xCCCCCC35b53c8a16404Ae414AFa31F30A5B35626",
     "0xb0dd7409",
     "0x00",
     "0x000000000000000000000000639fe6ab55c921f74e7fac1ee960c0b6293ba6120000000000000000000000000000000000000000000000000000000000003840000000000000000000000000e141425bc1594b8039de6390db1cdaf4397ea22b000000000000000000000000000000000000000000000000000000000001fa4000000000000000000000000035751007a407ca6feffe80b3cb397736d2cf4dbe"
   ],
   "0x35751007a407ca6FEFfE80b3cB397736D2cf4dbe": [
-    "0x74749a1c8b2faa03c7259F90919628b1241A2ea5",
+    "0xCCCCCC35b53c8a16404Ae414AFa31F30A5B35626",
     "0xb0dd7409",
     "0x00",
     "0x000000000000000000000000639fe6ab55c921f74e7fac1ee960c0b6293ba6120000000000000000000000000000000000000000000000000000000000003840000000000000000000000000b523ae262d20a936bc152e6023996e46fdc2a95d000000000000000000000000000000000000000000000000000000000001fa400000000000000000000000005979d7b546e38e414f7e9822514be443a4800529"
