+++ protocol/test/hardhat/utils/constants.js
@@ -59,5 +59,7 @@ module.exports = {
   L2_USDC: "0xaf88d065e77c8cC2239327C5EDb3A432268e5831",
   L2_USDT: "0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9",
 
+  L2_FERTILIZER: "0xFEFEFE2cfb089aEF0b0578573eF3CFAbC15f1490",
+
   ZERO_BYTES: "0x0000000000000000000000000000000000000000000000000000000000000000"
 };
